+++ common/src/main/java/whocraft/tardis_refined/constants/ModMessages.java
@@ -118,6 +118,7 @@ public class ModMessages {
     public static String ROOT_PLANT_CUT_OPEN = message("root_plant_cut_open");
     public static String WAYPOINT_LOADED = message("waypoint_loaded");
     public static String CANNOT_START_NO_FUEL = message("cannot_start_no_fuel");
+    public static String CANNOT_LAND = message("cannot_land");
     public static String CURRENT = message("current");
     public static String DESTINATION = message("destination");
     public static String SUBMIT = message("submit");